Folder restructurization
