+++ cvat-ui/src/containers/create-task-page/create-task-page.tsx
