+++ cvat-ui/src/containers/register-page/register-page.tsx
@@ -11,11 +11,11 @@ import {
     Modal,
 } from 'antd';
 
-import { registerAsync } from '../actions/auth-actions';
-import RegisterForm, { RegisterData } from '../components/register-form';
+import { registerAsync } from '../../actions/auth-actions';
+import RegisterForm, { RegisterData } from '../../components/register-page/register-form';
 
-import { CombinedState } from '../reducers/root-reducer';
-import { AuthState } from '../reducers/interfaces';
+import { CombinedState } from '../../reducers/root-reducer';
+import { AuthState } from '../../reducers/interfaces';
 
 interface StateToProps {
     auth: AuthState;
