+++ cvat-ui/src/containers/task-page/task-page.tsx
